Updated example read functions and renamed new cmlh5_to_xarray parser
